Move getroombyid controller tests out from matchup tests
